# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AllowedMentionsDto": {
        "additionalProperties": false,
        "properties": {
          "parse": {
            "description": "Types of mentions allowed",
            "example": [
              "users",
              "roles"
            ],
            "items": {
              "enum": [
                "everyone",
                "roles",
                "users"
              ],
              "type": "string"
            },
            "maxItems": 3,
            "type": "array"
          },
          "roles": {
            "description": "Discord role IDs to allow mentioning",
            "example": [
              "123456789012345678"
            ],
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "type": "array"
          },
          "users": {
            "description": "Discord user IDs to allow mentioning",
            "example": [
              "123456789012345678"
            ],
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "ChannelCategoryDto": {
        "additionalProperties": false,
        "properties": {
          "children": {
            "description": "Child channel IDs",
            "example": [
              "123456789012345678"
            ],
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "id": {
            "description": "Category ID",
            "example": "123456789012345678",
            "type": "string"
          },
          "name": {
            "description": "Category name",
            "example": "Text Channels",
            "type": "string"
          },
          "position": {
            "description": "Position",
            "example": 0,
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "position",
          "children"
        ],
        "type": "object"
      },
      "ChannelDetailResponseDto": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "description": "ISO 8601 creation timestamp",
            "example": "2025-01-15T14:30:00.000Z",
            "type": "string"
          },
          "id": {
            "description": "Discord channel ID",
            "example": "123456789012345678",
            "type": "string"
          },
          "lastMessageId": {
            "description": "Last message ID",
            "example": "123456789012345678",
            "nullable": true,
            "type": "string"
          },
          "name": {
            "description": "Channel name",
            "example": "general",
            "type": "string"
          },
          "nsfw": {
            "description": "Whether the channel is NSFW",
            "example": false,
            "type": "boolean"
          },
          "overwrites": {
            "description": "Per-target permission overwrites configured on the channel",
            "items": {
              "$ref": "#/components/schemas/ChannelPermissionOverwriteDto"
            },
            "type": "array"
          },
          "parentId": {
            "description": "Parent category ID",
            "example": "123456789012345678",
            "nullable": true,
            "type": "string"
          },
          "permissionOverwrites": {
            "description": "Whether the channel has any permission overwrites",
            "example": false,
            "type": "boolean"
          },
          "position": {
            "description": "Position",
            "example": 0,
            "type": "number"
          },
          "topic": {
            "description": "Channel topic",
            "example": "Discussion",
            "nullable": true,
            "type": "string"
          },
          "type": {
            "description": "Channel type",
            "example": "text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "position",
          "parentId",
          "topic",
          "nsfw",
          "lastMessageId",
          "createdAt",
          "permissionOverwrites",
          "overwrites"
        ],
        "type": "object"
      },
      "ChannelListResponseDto": {
        "additionalProperties": false,
        "properties": {
          "categories": {
            "items": {
              "$ref": "#/components/schemas/ChannelCategoryDto"
            },
            "type": "array"
          },
          "channels": {
            "items": {
              "$ref": "#/components/schemas/ChannelResponseDto"
            },
            "type": "array"
          }
        },
        "required": [
          "channels",
          "categories"
        ],
        "type": "object"
      },
      "ChannelPermissionOverwriteDto": {
        "additionalProperties": false,
        "properties": {
          "allow": {
            "description": "Allowed permission bits (Discord bitfield as a string)",
            "example": "1024",
            "type": "string"
          },
          "deny": {
            "description": "Denied permission bits (Discord bitfield as a string)",
            "example": "0",
            "type": "string"
          },
          "id": {
            "description": "Role or member ID the overwrite targets",
            "example": "123456789012345678",
            "type": "string"
          },
          "type": {
            "description": "Whether the overwrite targets a role or a member",
            "enum": [
              "role",
              "member"
            ],
            "example": "role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "allow",
          "deny"
        ],
        "type": "object"
      },
      "ChannelResponseDto": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "description": "ISO 8601 creation timestamp",
            "example": "2025-01-15T14:30:00.000Z",
            "type": "string"
          },
          "id": {
            "description": "Discord channel ID",
            "example": "123456789012345678",
            "type": "string"
          },
          "lastMessageId": {
            "description": "Last message ID",
            "example": "123456789012345678",
            "nullable": true,
            "type": "string"
          },
          "name": {
            "description": "Channel name",
            "example": "general",
            "type": "string"
          },
          "nsfw": {
            "description": "Whether the channel is NSFW",
            "example": false,
            "type": "boolean"
          },
          "parentId": {
            "description": "Parent category ID",
            "example": "123456789012345678",
            "nullable": true,
            "type": "string"
          },
          "permissionOverwrites": {
            "description": "Whether the channel has permission overwrites",
            "example": false,
            "type": "boolean"
          },
          "position": {
            "description": "Position in the channel list",
            "example": 0,
            "type": "number"
          },
          "topic": {
            "description": "Channel topic",
            "example": "Discussion",
            "nullable": true,
            "type": "string"
          },
          "type": {
            "description": "Channel type",
            "enum": [
              "text",
              "voice",
              "announcement",
              "category"
            ],
            "example": "text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "position",
          "parentId",
          "topic",
          "nsfw",
          "permissionOverwrites"
        ],
        "type": "object"
      },
...
      "GetMessagesResponseDto": {
        "additionalProperties": false,
        "properties": {
          "hasMore": {
            "description": "Whether there are more messages available",
            "example": false,
            "type": "boolean"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageDto"
            },
            "type": "array"
          }
        },
        "required": [
          "messages",
          "hasMore"
        ],
        "type": "object"
      },
...
      "MessageDto": {
        "additionalProperties": false,
        "properties": {
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/MessageAttachmentDto"
            },
            "type": "array"
          },
          "author": {
            "$ref": "#/components/schemas/MessageAuthorDto"
          },
          "content": {
            "description": "Message content",
            "example": "Hello!",
            "type": "string"
          },
          "embeds": {
            "items": {
              "$ref": "#/components/schemas/MessageEmbedDto"
            },
            "type": "array"
          },
          "id": {
            "description": "Discord message ID",
            "example": "123456789012345678",
            "type": "string"
          },
          "mentions": {
            "items": {
              "$ref": "#/components/schemas/MentionDto"
            },
            "type": "array"
          },
          "timestamp": {
            "description": "ISO 8601 timestamp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "content",
          "author",
          "timestamp",
          "embeds",
          "attachments",
          "mentions"
        ],
        "type": "object"
      },
...
      "SendMessageDto": {
        "additionalProperties": false,
        "properties": {
          "allowed_mentions": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AllowedMentionsDto"
              }
            ],
            "description": "Allowed mentions configuration"
          },
          "content": {
            "description": "Message content (max 2000 chars, optional if embeds provided)",
            "example": "Hello from MCDI!",
            "maxLength": 2000,
            "type": "string"
          },
          "embeds": {
            "description": "Embed objects (max 10)",
            "items": {
              "$ref": "#/components/schemas/EmbedDto"
            },
            "maxItems": 10,
            "type": "array"
          },
          "tts": {
            "default": false,
            "description": "Whether the message is text-to-speech",
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "SendMessageResponseDto": {
        "additionalProperties": false,
        "properties": {
          "author": {
            "$ref": "#/components/schemas/MessageAuthorDto"
          },
          "channelId": {
            "description": "Discord channel ID",
            "example": "123456789012345678",
            "type": "string"
          },
          "content": {
            "description": "Message content",
            "example": "Hello!",
            "type": "string"
          },
          "id": {
            "description": "Discord message ID",
            "example": "123456789012345678",
            "type": "string"
          },
          "timestamp": {
            "description": "ISO 8601 timestamp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "channelId",
          "content",
          "timestamp",
          "author"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "api-key": {
        "description": "Project API key (prefix.secret format)",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "session-token": {
        "bearerFormat": "Token",
        "description": "Enter your session token (from admin login or POST /auth/validate)",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "MicroClub Discord Identity API.\n\n**MC Project endpoints** authenticate with `X-API-Key` header.\n\n**Admin endpoints** authenticate with `Authorization: Bearer <token>`.\n\nScopes and operations are granted per project–server pair by a system administrator.",
    "title": "MCDI API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/servers/{serverId}/channels": {
      "get": {
        "description": "Returns all channels the bot can access in the server, with categories.",
        "operationId": "AdminChannelsController_listChannels",
        "parameters": [
          {
            "description": "Discord server ID",
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          },
          {
            "description": "Filter by channel type",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "text",
                "voice",
                "announcement",
                "category"
              ],
              "type": "string"
            }
          },
          {
            "description": "Filter by parent category ID",
            "in": "query",
            "name": "categoryId",
            "required": false,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelListResponseDto"
                }
              }
            },
            "description": "List of channels and categories."
          },
          "400": {
            "description": "Invalid query parameters."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "System Admin access required."
          },
          "404": {
            "description": "Server not found or bot not connected."
          }
        },
        "security": [
          {
            "session-token": []
          }
        ],
        "summary": "List all channels in a server (admin session)",
        "tags": [
          "Channels"
        ]
      }
    },
    "/api/admin/servers/{serverId}/channels/{channelId}": {
      "get": {
        "description": "Returns detailed information about a specific channel, including its permission overwrites.",
        "operationId": "AdminChannelsController_getChannel",
        "parameters": [
          {
            "description": "Discord server ID",
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          },
          {
            "description": "Discord channel ID",
            "in": "path",
            "name": "channelId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelDetailResponseDto"
                }
              }
            },
            "description": "Channel details."
          },
          "400": {
            "description": "Invalid channel ID format."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "System Admin access required."
          },
          "404": {
            "description": "Channel does not exist or bot lacks access."
          }
        },
        "security": [
          {
            "session-token": []
          }
        ],
        "summary": "Get channel details (admin session)",
        "tags": [
          "Channels"
        ]
      }
    },
    "/api/admin/servers/{serverId}/channels/{channelId}/messages": {
      "get": {
        "description": "Retrieves message history from a text channel. Read-only; supports pagination and author filtering.",
        "operationId": "AdminChannelsController_getMessages",
        "parameters": [
          {
            "description": "Discord server ID",
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          },
          {
            "description": "Discord channel ID",
            "in": "path",
            "name": "channelId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          },
          {
            "description": "Max number of messages (1-100)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "example": 50,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Get messages before this Discord snowflake ID",
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "example": "123456789012345678",
              "pattern": "/^\\d{17,20}$/",
              "type": "string"
            }
          },
          {
            "description": "Get messages after this Discord snowflake ID",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "example": "123456789012345678",
              "pattern": "/^\\d{17,20}$/",
              "type": "string"
            }
          },
          {
            "description": "Filter by author Discord ID",
            "in": "query",
            "name": "authorId",
            "required": false,
            "schema": {
              "example": "123456789012345678",
              "pattern": "/^\\d{17,20}$/",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMessagesResponseDto"
                }
              }
            },
            "description": "List of messages."
          },
          "400": {
            "description": "Invalid query parameters."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "System Admin access required."
          },
          "404": {
            "description": "Channel does not exist or bot lacks access."
          },
          "429": {
            "description": "Rate limit exceeded. 30 requests per minute."
          }
        },
        "security": [
          {
            "session-token": []
          }
        ],
        "summary": "Get recent messages from a channel (admin session)",
        "tags": [
          "Channels"
        ]
      }
    },
...
    "/api/servers/{serverId}/channels": {
      "get": {
        "description": "Returns all channels the bot can access in the server, with categories.",
        "operationId": "ChannelsController_listChannels",
        "parameters": [
          {
            "description": "Discord server ID",
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          },
          {
            "description": "Filter by channel type",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "text",
                "voice",
                "announcement",
                "category"
              ],
              "type": "string"
            }
          },
          {
            "description": "Filter by parent category ID",
            "in": "query",
            "name": "categoryId",
            "required": false,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelListResponseDto"
                }
              }
            },
            "description": "List of channels and categories."
          },
          "400": {
            "description": "Invalid query parameters."
          },
          "401": {
            "description": "Missing or invalid API key."
          },
          "403": {
            "description": "Project does not have READ access to this server."
          },
          "404": {
            "description": "Server not found or bot not connected."
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "List all channels in a server",
        "tags": [
          "Channels"
        ]
      }
    },
    "/api/servers/{serverId}/channels/{channelId}": {
      "get": {
        "description": "Returns detailed information about a specific channel.",
        "operationId": "ChannelsController_getChannel",
        "parameters": [
          {
            "description": "Discord server ID",
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          },
          {
            "description": "Discord channel ID",
            "in": "path",
            "name": "channelId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelDetailResponseDto"
                }
              }
            },
            "description": "Channel details."
          },
          "400": {
            "description": "Invalid channel ID format."
          },
          "401": {
            "description": "Missing or invalid API key."
          },
          "403": {
            "description": "Project does not have READ access to this server."
          },
          "404": {
            "description": "Channel does not exist or bot lacks access."
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Get channel details",
        "tags": [
          "Channels"
        ]
      }
    },
    "/api/servers/{serverId}/channels/{channelId}/messages": {
      "get": {
        "description": "Retrieves message history from a text channel. Supports pagination and author filtering.",
        "operationId": "ChannelsController_getMessages",
        "parameters": [
          {
            "description": "Discord server ID",
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          },
          {
            "description": "Discord channel ID",
            "in": "path",
            "name": "channelId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          },
          {
            "description": "Max number of messages (1-100)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "example": 50,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Get messages before this Discord snowflake ID",
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "example": "123456789012345678",
              "pattern": "/^\\d{17,20}$/",
              "type": "string"
            }
          },
          {
            "description": "Get messages after this Discord snowflake ID",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "example": "123456789012345678",
              "pattern": "/^\\d{17,20}$/",
              "type": "string"
            }
          },
          {
            "description": "Filter by author Discord ID",
            "in": "query",
            "name": "authorId",
            "required": false,
            "schema": {
              "example": "123456789012345678",
              "pattern": "/^\\d{17,20}$/",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMessagesResponseDto"
                }
              }
            },
            "description": "List of messages."
          },
          "400": {
            "description": "Invalid query parameters."
          },
          "401": {
            "description": "Missing or invalid API key."
          },
          "403": {
            "description": "Project does not have READ access to this server."
          },
          "404": {
            "description": "Channel does not exist or bot lacks access."
          },
          "429": {
            "description": "Rate limit exceeded. 10 requests per minute per project."
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Get recent messages from a channel",
        "tags": [
          "Channels"
        ]
      },
      "post": {
        "description": "Sends a message with optional embeds and mention configuration.",
        "operationId": "ChannelsController_sendMessage",
        "parameters": [
          {
            "description": "Discord server ID",
            "in": "path",
            "name": "serverId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          },
          {
            "description": "Discord channel ID",
            "in": "path",
            "name": "channelId",
            "required": true,
            "schema": {
              "example": "123456789012345678",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendMessageResponseDto"
                }
              }
            },
            "description": "Message sent successfully."
          },
          "400": {
            "description": "Invalid message content or embed validation failed."
          },
          "401": {
            "description": "Missing or invalid API key."
          },
          "403": {
            "description": "Project does not have SEND_MESSAGES operation on this server."
          },
          "404": {
            "description": "Channel does not exist or bot lacks access."
          },
          "429": {
            "description": "Rate limit exceeded. 5 messages per minute per project."
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Send a message to a Discord channel",
        "tags": [
          "Channels"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "API Server",
      "url": "http://localhost:3000"
    }
  ],
...
}